+++ src/Billable.php
@@ -13,19 +13,13 @@ use Stripe\Invoice as StripeInvoice;
 use Stripe\Customer as StripeCustomer;
 use Stripe\BankAccount as StripeBankAccount;
 use Stripe\InvoiceItem as StripeInvoiceItem;
+use Laravel\Cashier\Exceptions\InvalidStripeCustomer;
 use Stripe\Error\InvalidRequest as StripeErrorInvalidRequest;
 use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 use Symfony\Component\HttpKernel\Exception\AccessDeniedHttpException;
 
 trait Billable
 {
-    /**
-     * The Stripe API key.
-     *
-     * @var string
-     */
-    protected static $stripeKey;
-
     /**
      * Make a "one off" charge on the customer for the given amount.
      *
@@ -42,7 +36,7 @@ trait Billable
 
         $options['amount'] = $amount;
 
-        if (!array_key_exists('source', $options) && $this->stripe_id) {
+        if ($this->stripe_id) {
             $options['customer'] = $this->stripe_id;
         }
 
@@ -50,7 +44,7 @@ trait Billable
             throw new InvalidArgumentException('No payment source provided.');
         }
 
-        return StripeCharge::create($options, ['api_key' => $this->getStripeKey()]);
+        return StripeCharge::create($options, Cashier::stripeOptions());
     }
 
     /**
@@ -65,7 +59,7 @@ trait Billable
     {
         $options['charge'] = $charge;
 
-        return StripeRefund::create($options, ['api_key' => $this->getStripeKey()]);
+        return StripeRefund::create($options, Cashier::stripeOptions());
     }
 
     /**
@@ -89,9 +83,7 @@ trait Billable
      */
     public function tab($description, $amount, array $options = [])
     {
-        if (!$this->stripe_id) {
-            throw new InvalidArgumentException(class_basename($this) . ' is not a Stripe customer. See the createAsStripeCustomer method.');
-        }
+        $this->assertCustomerExists();
 
         $options = array_merge([
             'customer' => $this->stripe_id,
@@ -100,7 +92,7 @@ trait Billable
             'description' => $description,
         ], $options);
 
-        return StripeInvoiceItem::create($options, ['api_key' => $this->getStripeKey()]);
+        return StripeInvoiceItem::create($options, Cashier::stripeOptions());
     }
 
     /**
@@ -227,17 +219,15 @@ trait Billable
      */
     public function invoice(array $options = [])
     {
-        if ($this->stripe_id) {
-            $parameters = array_merge($options, ['customer' => $this->stripe_id]);
+        $this->assertCustomerExists();
 
-            try {
-                return StripeInvoice::create($parameters, $this->getStripeKey())->pay();
-            } catch (StripeErrorInvalidRequest $e) {
-                return false;
-            }
+        $parameters = array_merge($options, ['customer' => $this->stripe_id]);
+
+        try {
+            return StripeInvoice::create($parameters, Cashier::stripeOptions())->pay();
+        } catch (StripeErrorInvalidRequest $e) {
+            return false;
         }
-
-        return true;
     }
 
     /**
@@ -247,11 +237,10 @@ trait Billable
      */
     public function upcomingInvoice()
     {
+        $this->assertCustomerExists();
+
         try {
-            $stripeInvoice = StripeInvoice::upcoming(
-                ['customer' => $this->stripe_id],
-                ['api_key' => $this->getStripeKey()]
-            );
+            $stripeInvoice = StripeInvoice::upcoming(['customer' => $this->stripe_id], Cashier::stripeOptions());
 
             return new Invoice($this, $stripeInvoice);
         } catch (StripeErrorInvalidRequest $e) {
@@ -269,13 +258,12 @@ trait Billable
     {
         try {
             $stripeInvoice = StripeInvoice::retrieve(
-                $id,
-                $this->getStripeKey()
+                $id, Cashier::stripeOptions()
             );
 
-            $stripeInvoice->lines = StripeInvoice::retrieve($id, $this->getStripeKey())
-                ->lines
-                ->all(['limit' => 1000]);
+            $stripeInvoice->lines = StripeInvoice::retrieve($id, Cashier::stripeOptions())
+                        ->lines
+                        ->all(['limit' => 1000]);
 
             return new Invoice($this, $stripeInvoice);
         } catch (Exception $e) {
@@ -284,7 +272,7 @@ trait Billable
     }
 
     /**
-     * Find an invoice or throw a 404 error.
+     * Find an invoice or throw a 404 or 403 error.
      *
      * @param  string  $id
      * @return \Laravel\Cashier\Invoice
@@ -325,6 +313,8 @@ trait Billable
      */
     public function invoices($includePending = false, $parameters = [])
     {
+        $this->assertCustomerExists();
+
         $invoices = [];
 
         $parameters = array_merge(['limit' => 24], $parameters);
@@ -364,6 +354,8 @@ trait Billable
      */
     public function cards($parameters = [])
     {
+        $this->assertCustomerExists();
+
         $cards = [];
 
         $parameters = array_merge(['limit' => 24], $parameters);
@@ -409,9 +401,11 @@ trait Billable
      */
     public function updateCard($token)
     {
+        $this->assertCustomerExists();
+
         $customer = $this->asStripeCustomer();
 
-        $token = StripeToken::retrieve($token, ['api_key' => $this->getStripeKey()]);
+        $token = StripeToken::retrieve($token, Cashier::stripeOptions());
 
         // If the given token already has the card as their default source, we can just
         // bail out of the method now. We don't need to keep adding the same card to
@@ -500,6 +494,8 @@ trait Billable
      */
     public function applyCoupon($coupon)
     {
+        $this->assertCustomerExists();
+
         $customer = $this->asStripeCustomer();
 
         $customer->coupon = $coupon;
@@ -562,6 +558,20 @@ trait Billable
     }
 
     /**
+     * Determine if the entity has a Stripe customer ID and throw an exception if not.
+     *
+     * @return void
+     *
+     * @throws \Laravel\Cashier\Exceptions\InvalidStripeCustomer
+     */
+    protected function assertCustomerExists()
+    {
+        if (! $this->stripe_id) {
+            throw InvalidStripeCustomer::nonCustomer($this);
+        }
+    }
+
+    /**
      * Create a Stripe customer for the given model.
      *
      * @param  array  $options
@@ -577,8 +587,7 @@ trait Billable
         // user from Stripe. This ID will correspond with the Stripe user instances
         // and allow us to retrieve users from Stripe later when we need to work.
         $customer = StripeCustomer::create(
-            $options,
-            $this->getStripeKey()
+            $options, Cashier::stripeOptions()
         );
 
         $this->stripe_id = $customer->id;
@@ -596,13 +605,24 @@ trait Billable
      */
     public function updateStripeCustomer(array $options = [])
     {
-        $customer = StripeCustomer::update(
-            $this->stripe_id,
-            $options,
-            $this->getStripeKey()
+        return StripeCustomer::update(
+            $this->stripe_id, $options, Cashier::stripeOptions()
         );
+    }
 
-        return $customer;
+    /**
+     * Get the Stripe customer instance for the current user and token.
+     *
+     * @param  array  $options
+     * @return \Stripe\Customer
+     */
+    public function createOrGetStripeCustomer(array $options = [])
+    {
+        if ($this->stripe_id) {
+            return $this->asStripeCustomer();
+        }
+
+        return $this->createAsStripeCustomer($options);
     }
 
     /**
@@ -612,7 +632,7 @@ trait Billable
      */
     public function asStripeCustomer()
     {
-        return StripeCustomer::retrieve($this->stripe_id, $this->getStripeKey());
+        return StripeCustomer::retrieve($this->stripe_id, Cashier::stripeOptions());
     }
 
     /**
@@ -622,7 +642,7 @@ trait Billable
      */
     public function preferredCurrency()
     {
-        return Cashier::usesCurrency();
+        return config('cashier.currency');
     }
 
     /**
@@ -633,35 +653,6 @@ trait Billable
     public function taxPercentage()
     {
         return 0;
-    }
-
-    /**
-     * Get the Stripe API key.
-     *
-     * @return string
-     */
-    public static function getStripeKey()
-    {
-        if (static::$stripeKey) {
-            return static::$stripeKey;
-        }
-
-        if ($key = getenv('STRIPE_SECRET')) {
-            return $key;
-        }
-
-        return config('services.stripe.secret');
-    }
-
-    /**
-     * Set the Stripe API key.
-     *
-     * @param  string  $key
-     * @return void
-     */
-    public static function setStripeKey($key)
-    {
-        static::$stripeKey = $key;
     }
 
     /**